Add Grantlee5 requirement to CMakeLists and convert Int to Long. (#4)
* Add Grantlee5 requirement to CMakeLists and convert Int to Long.

* Update storages.cpp

* Update storagepool.h

* Update storagepool.cpp

The better way as it was and renamed.

* Update CMakeLists.txt

+++ src/lib/storagepool.h
@@ -66,7 +66,7 @@ public:
 
     bool build(int flags);
     bool create(int flags);
-    StorageVol *createStorageVolume(const QString &name, const QString &format, int sizeGiB, int flags);
+    StorageVol *createStorageVolume(const QString &name, const QString &format, qint64 sizeGiB, int flags);
 //    bool cloneStorageVolume(StorageVol *volume, const QString &name, const QString &format, int flags);
     StorageVol *getVolume(const QString &name);
